functions/imageprocess/requirements.txt
# Function dependencies, for example:
# package>=version
pillow
google-cloud-bigquery
google-cloud-vision
